Handled the exceptions about the function of Comment

+++ src/main/java/cn/lnsf/community/controller/QuestionController.java
@@ -20,7 +20,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String question(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             Model model) {
 
         QuestionDTO questionDTO = questionService.getById(id);
